+++ lvlspy/calculate/__init__.py
@@ -1,0 +1,6 @@
+"""
+A subdirectoy of lvlspy to handle calculations
+"""
+
+from lvlspy.calculate.weisskopf import *
+from lvlspy.calculate.evolve import *